@@ -1,32 +1,29 @@
 {
   "name": "mastermind",
-  "version": "0.0.1",
   "license": "MIT",
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "build": "NODE_ENV=production webpack",
-    "dev": "NODE_ENV=development webpack serve",
-    "prettier": "prettier --write *.js .prettierrc *.json src/*"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "check": "svelte-check --tsconfig ./tsconfig.app.json && tsc -p tsconfig.node.json",
+    "format": "prettier --write 'src/**/*.svelte' '*.ts' '*.json' '*.js'"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.20",
-    "css-loader": "^7.1.2",
-    "css-minimizer-webpack-plugin": "^7.0.0",
-    "mini-css-extract-plugin": "^2.9.1",
-    "postcss": "^8.4.47",
-    "postcss-loader": "^8.1.1",
-    "prettier": "^3.3.3",
-    "prettier-plugin-svelte": "^3.2.6",
-    "style-loader": "^4.0.0",
-    "svelte": "^4.2.19",
-    "svelte-loader": "^3.2.3",
-    "svelte-modals": "^1.3.0",
-    "svelte-preprocess": "^6.0.2",
-    "terser-webpack-plugin": "^5.3.10",
-    "webpack": "^5.95.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.1.0"
+    "@sveltejs/vite-plugin-svelte": "^6.2.1",
+    "@tsconfig/svelte": "^5.0.6",
+    "@types/node": "^24.10.1",
+    "prettier": "^3.7.1",
+    "prettier-plugin-svelte": "^3.4.0",
+    "svelte": "^5.43.8",
+    "svelte-check": "^4.3.4",
+    "typescript": "~5.9.3",
+    "vite": "^7.2.4"
   },
   "dependencies": {
-    "tailwindcss": "^3.4.13"
+    "@tailwindcss/vite": "^4.1.17",
+    "svelte-modals": "^2.0.1",
+    "tailwindcss": "^4.1.17"
   }
 }
