@@ -1,14 +1,31 @@
-<script>
-  export let color
+<script lang="ts">
+  import type { BlackOrWhite } from "./types"
 
-  $: filled = color == "white" || color == "black"
+  const { color = "" } = $props<{
+    color: BlackOrWhite | ""
+  }>()
+
+  const bg = $derived.by(() => {
+    switch (color) {
+      case "white":
+        return "bg-white"
+      case "black":
+        return "bg-black"
+      default:
+        return "bg-slate-200 dark:bg-slate-700"
+    }
+  })
+
+  const filled = $derived(color !== "")
 </script>
 
 <div
-  class="block h-1 w-1 rounded-full shadow"
-  class:peg={true}
-  class:h-3={filled}
-  class:w-3={filled}
-  class:bg-black={color == "black"}
-  class:bg-white={color == "white"}
-/>
+  class={[
+    "block",
+    "rounded-full",
+    "shadow",
+    "transition-colors",
+    bg,
+    filled ? "size-3" : "size-1",
+  ]}
+></div>
