@@ -1,27 +1,36 @@
-<script>
+<script lang="ts">
+  import { onMount } from "svelte"
+
   import Peg from "./Peg.svelte"
-  import { createEventDispatcher, onMount } from "svelte"
+
   import { toggleColor } from "./utils.js"
+  import { Color, type Answer } from "./types"
 
-  export let canGuess
+  let { canGuess = true, submitGuess } = $props<{
+    canGuess: boolean
+    submitGuess: (answer: Answer) => Promise<void>
+  }>()
 
-  let dispatch = createEventDispatcher()
-  let guess = initialGuess()
+  let guess = $state<Answer>(initialGuess())
 
-  function initialGuess() {
-    return ["", "", "", ""]
+  function initialGuess(): Answer {
+    return Array(4).fill(Color.Empty) as Answer
   }
 
   onMount(() => {
     document.addEventListener("keyup", handleKeyUp)
   })
 
-  function handleSubmit() {
-    dispatch("guess", { guess: guess })
+  async function handleSubmit() {
+    await submitGuess(guess)
     guess = initialGuess()
   }
 
-  function handleKeyUp(evt) {
+  function handleKeyUp(evt: KeyboardEvent) {
+    if (!canGuess) {
+      return
+    }
+
     switch (evt.key) {
       case "1":
       case "a":
@@ -56,69 +65,58 @@
     }
   }
 
-  $: submittable = canGuess && guess.filter((s) => s == "").length == 0
+  let submittable = $derived(
+    canGuess && guess.filter((s) => s == "").length == 0,
+  )
 </script>
 
-<div />
+<div></div>
 
-{#each guess as peg}
+{#each guess as peg, i}
   <div>
-    <Peg bind:value={peg} lg />
+    <Peg bind:value={guess[i]} lg />
   </div>
 {/each}
 
 <div>
   <button
-    on:click={handleSubmit}
+    onclick={handleSubmit}
     disabled={!submittable}
-    class="block w-10 h-10 p-2 bg-blue-500 text-white rounded-md disabled:bg-slate-400 shadow-lg disabled:shadow transition-all"
+    class="block size-10 p-2 bg-blue-500 text-white rounded-md disabled:bg-slate-400 shadow-lg disabled:shadow transition-all cursor-pointer fill-current"
+    title="Submit guess"
   >
-    <svg
-      aria-hidden="true"
-      focusable="false"
-      data-prefix="far"
-      data-icon="check"
-      class="svg-inline--fa fa-check fa-w-16"
-      role="img"
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 512 512"
-    >
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 640 640">
       <path
-        fill="currentColor"
-        d="M435.848 83.466L172.804
-          346.51l-96.652-96.652c-4.686-4.686-12.284-4.686-16.971 0l-28.284
-          28.284c-4.686 4.686-4.686 12.284 0 16.971l133.421 133.421c4.686 4.686
-          12.284 4.686 16.971 0l299.813-299.813c4.686-4.686 4.686-12.284
-          0-16.971l-28.284-28.284c-4.686-4.686-12.284-4.686-16.97 0z"
+        d="M530.8 134.1C545.1 144.5 548.3 164.5 537.9 178.8L281.9 530.8C276.4 538.4 267.9 543.1 258.5 543.9C249.1 544.7 240 541.2 233.4 534.6L105.4 406.6C92.9 394.1 92.9 373.8 105.4 361.3C117.9 348.8 138.2 348.8 150.7 361.3L252.2 462.8L486.2 141.1C496.6 126.8 516.6 123.6 530.9 134z"
       />
     </svg>
   </button>
 </div>
 
-<div />
+<div></div>
 
 <div
-  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono w-6 h-6 select-none"
+  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono size-8 p-1 select-none"
 >
   1
 </div>
 <div
-  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono w-6 h-6 select-none"
+  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono size-8 p-1 select-none"
 >
   2
 </div>
 <div
-  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono w-6 h-6 select-none"
+  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono size-8 p-1 select-none"
 >
   3
 </div>
 <div
-  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono w-6 h-6 select-none"
+  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono size-8 p-1 select-none"
 >
   4
 </div>
 <div
-  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono w-6 h-6 select-none"
+  class="grid place-items-center border border-slate-300 dark:border-slate-600 text-slate-400 dark:text-slate-500 transition-colors rounded text-sm font-mono size-8 p-1 select-none"
 >
   ⮐
 </div>
