@@ -1,13 +1,14 @@
-<script>
+<script lang="ts">
   import { onMount, onDestroy } from "svelte"
-  import { closeModal } from "svelte-modals"
   import Peg from "./Peg.svelte"
+  import type { Answer } from "./types"
 
-  export let isOpen
-
-  export let answer
-  export let guesses
-  export let onConfirm
+  const { isOpen, close, answer, guesses } = $props<{
+    isOpen: boolean
+    close: () => void
+    answer?: Answer
+    guesses?: number
+  }>()
 
   onMount(() => {
     document.addEventListener("keyup", handleKeyUp)
@@ -17,19 +18,14 @@
     document.removeEventListener("keyup", handleKeyUp)
   })
 
-  function handleConfirm() {
-    onConfirm()
-    closeModal()
-  }
-
-  function handleKeyUp(evt) {
+  function handleKeyUp(evt: KeyboardEvent) {
     switch (evt.key) {
       case "Enter":
-        handleConfirm()
+        close()
         break
 
       case "Escape":
-        closeModal()
+        close()
         break
     }
   }
@@ -38,12 +34,14 @@
 {#if isOpen}
   <div
     role="dialog"
+    aria-modal="true"
+    aria-labelledby="modal-title"
     class="fixed inset-0 flex justify-center items-center pointer-events-none z-20"
   >
     <div
       class="min-w-[20rem] max-w-[40ch] rounded-md bg-slate-100 dark:bg-slate-700 p-6 z-20 pointer-events-auto"
     >
-      <h2 class="text-xl font-bold mb-2">You won! 🎉</h2>
+      <h2 class="text-xl font-bold mb-2" id="modal-title">You won! 🎉</h2>
       <p class="my-4">The answer was:</p>
 
       <div class="flex flex-row gap-4 my-4 place-content-center">
@@ -59,7 +57,7 @@
       <div class="mt-6 text-center">
         <button
           type="button"
-          on:click={handleConfirm}
+          onclick={close}
           class="bg-blue-500 text-white rounded py-2 px-4 pointer">OK</button
         >
       </div>
