@@ -1,14 +1,7 @@
-<div />
-<div
-  class="block w-1 h-1 m-5 rounded-full bg-slate-300 dark:bg-slate-600 transition-colors "
-/>
-<div
-  class="block w-1 h-1 m-5 rounded-full bg-slate-300 dark:bg-slate-600 transition-colors"
-/>
-<div
-  class="block w-1 h-1 m-5 rounded-full bg-slate-300 dark:bg-slate-600 transition-colors"
-/>
-<div
-  class="block w-1 h-1 m-5 rounded-full bg-slate-300 dark:bg-slate-600 transition-colors"
-/>
-<div />
+<div></div>
+{#each Array(4).fill(0)}
+  <div
+    class="block size-1 m-5 rounded-full bg-slate-300 dark:bg-slate-600 transition-colors"
+  ></div>
+{/each}
+<div></div>
